command_base64: Move from base.py to base64.py
